fix(sqlite): reescribir queries con prepared statements y hardening Electron
