+++ apps/electron/main.js
@@ -58,7 +58,7 @@ function createWindow() {
       preload:          path.join(__dirname, 'preload.js'),
       contextIsolation: true,
       nodeIntegration:  false,
-      sandbox:          false,
+      sandbox:          true,
       webSecurity:      !IS_DEV,
     },
   });
